database_adapter/requirements.txt
mysql-connector-python
python-dotenv
flask-restful